Check to see whether sssd.conf was created before trying to reinstall ipaclient

+++ manifests/clientinstall.pp
@@ -30,19 +30,26 @@ define ipa::clientinstall (
     default => ''
   }
 
-  $clientinstallcmd = shellquote('/usr/sbin/ipa-client-install',"--server=${masterfqdn}","--hostname=${host}","--domain=${domain}","--realm=${realm}","--password=${otp}",$mkhomediropt,$ntpopt,$fixedprimaryopt,'--unattended')
-  $dc = prefix([regsubst($domain,'(\.)',',dc=','G')],'dc=')
+  ## the plugin helps or tries to mitigate against lost of network connectivity
+  #  with ipa master.  The logic here is if sssd.conf is not present then 
+  #  ipa-client-install failed.  Therefore /etc/ipa/ca.crt must be removed if 
+  #  ipa-client-install is to be run again
 
-  exec { "client-install-${host}":
-    command   => "/bin/echo | rm -f /etc/ipa/ca.crt && ${clientinstallcmd}",
-    unless    => shellquote('/bin/bash','-c',"LDAPTLS_REQCERT=never /usr/bin/ldapsearch -LLL -x -H ldaps://${masterfqdn} -D uid=admin,cn=users,cn=accounts,${dc} -b ${dc} -w ${adminpw} fqdn=${host} | /bin/grep ^krbLastPwdChange"),
-    timeout   => '0',
-    tries     => '60',
-    try_sleep => '90',
-    returns   => ['0','1'],
-    logoutput => 'on_failure'
-  }
+  unless $::ipa_clientinstall 
+    $clientinstallcmd = shellquote('/usr/sbin/ipa-client-install',"--server=${masterfqdn}","--hostname=${host}","--domain=${domain}","--realm=${realm}","--password=${otp}",$mkhomediropt,$ntpopt,$fixedprimaryopt,'--unattended')
+    $dc = prefix([regsubst($domain,'(\.)',',dc=','G')],'dc=')
 
-  ipa::flushcache { "client-${host}":
+    exec { "client-install-${host}":
+      command   => "/bin/echo | rm -f /etc/ipa/ca.crt && ${clientinstallcmd}",
+      unless    => shellquote('/bin/bash','-c',"LDAPTLS_REQCERT=never /usr/bin/ldapsearch -LLL -x -H ldaps://${masterfqdn} -D uid=admin,cn=users,cn=accounts,${dc} -b ${dc} -w ${adminpw} fqdn=${host} | /bin/grep ^krbLastPwdChange"),
+      timeout   => '0',
+      tries     => '5',
+      try_sleep => '10',
+      returns   => ['0','1'],
+      logoutput => 'on_failure'
+    }
+
+    ipa::flushcache { "client-${host}":
+    }
   }
 }
